requirements.txt
flask
azure-storage
azure-mgmt-resource
azure-mgmt-media
